# transactions/swagger/transaction.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Transactions",
    "description": "API Transactions",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:8003",
      "description": "API Transactions"
    }
  ],
  "paths": {
    "/transactions": {
      "post": {
        "summary": "Create new transaction",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Transaction"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transaction successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          }
        }
      }
    },
    "/transactions/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "schema": {
            "type": "string"
          },
          "required": true
        }
      ],
      "get": {
        "summary": "Get transaction by ID",
        "responses": {
          "200": {
            "description": "Transaction successfully received",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionById"
                }
              }
            }
          },
          "404": {
            "description": "Transaction not found",
            "content": {
              "application/json": {
                "example": "Not Found"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Transaction": {
        "type": "object",
        "properties": {
          "transactionValue": {
            "type": "number"
          },
          "cardHolder": {
            "type": "string"
          },
          "cardNumber": {
            "type": "string"
          },
          "cardExp": {
            "type": "string"
          },
          "cardCvv": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "TransactionById": {
        "type": "object",
        "properties": {
          "transactionId": {
            "type": "string"
          },
          "clientId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      }
    }
  }
}